+++ ui/main/game/galactic_war/shared/js/gw_faction_1.js
@@ -291,19 +291,14 @@ define(function () {
     },
     {
       name: "Vidicarus",
-      character: "!LOC:Relentless",
+      character: "!LOC:Factory",
       color: [
         [255, 153, 204],
         [192, 192, 192],
       ],
       personality: {
-        metal_drain_check: 0.44,
-        energy_drain_check: 0.55,
-        metal_demand_check: 0.61,
-        energy_demand_check: 0.7,
-        neural_data_mod: 1.2,
-        adv_eco_mod: 1.2,
-        adv_eco_mod_alone: 0.95,
+        metal_demand_check: 0.99,
+        energy_demand_check: 0.99,
       },
       commander:
         "/pa/units/commanders/imperial_vidicarus/imperial_vidicarus.json",