+++ ui/main/game/galactic_war/shared/js/gw_faction_3.js
@@ -288,19 +288,14 @@ define(function () {
     },
     {
       name: "XOV",
-      character: "!LOC:Relentless",
+      character: "!LOC:Factory",
       color: [
         [255, 255, 0],
         [192, 192, 192],
       ],
       personality: {
-        metal_drain_check: 0.44,
-        energy_drain_check: 0.55,
-        metal_demand_check: 0.61,
-        energy_demand_check: 0.7,
-        neural_data_mod: 1.2,
-        adv_eco_mod: 1.2,
-        adv_eco_mod_alone: 0.95,
+        metal_demand_check: 0.99,
+        energy_demand_check: 0.99,
       },
       commander: "/pa/units/commanders/raptor_xov/raptor_xov.json",
     },